+++ lib/any_cache/adapters.rb
@@ -13,6 +13,7 @@ module AnyCache::Adapters
   require_relative 'adapters/active_support_redis_cache_store'
   require_relative 'adapters/active_support_memory_store'
   require_relative 'adapters/active_support_mem_cache_store'
+  require_relative 'adapters/active_support_dalli_store'
 
   class << self
     # @param driver [Object]
@@ -22,7 +23,7 @@ module AnyCache::Adapters
     #
     # @api private
     # @since 0.1.0
-    # rubocop:disable Metrics/LineLength
+    # rubocop:disable Metrics/LineLength, Metrics/AbcSize
     def build(driver)
       case
       when RedisStore.supported_driver?(driver)                   then RedisStore.new(driver)
@@ -32,11 +33,12 @@ module AnyCache::Adapters
       when ActiveSupportMemoryStore.supported_driver?(driver)     then ActiveSupportMemoryStore.new(driver)
       when ActiveSupportFileStore.supported_driver?(driver)       then ActiveSupportFileStore.new(driver)
       when ActiveSupportMemCacheStore.supported_driver?(driver)   then ActiveSupportMemCacheStore.new(driver)
+      when ActiveSupportDalliStore.supported_driver?(driver)      then ActiveSupportDalliStore.new(driver)
       when Delegator.supported_driver?(driver)                    then Delegator.new(driver)
       else
         raise AnyCache::UnsupportedDriverError
       end
     end
-    # rubocop:enable Metrics/LineLength
+    # rubocop:enable Metrics/LineLength, Metrics/AbcSize
   end
 end