+++ spec/support/spec_support/cache.rb
@@ -8,4 +8,5 @@ module SpecSupport::Cache
   require_relative 'cache/active_support_memory_store'
   require_relative 'cache/active_support_redis_cache_store'
   require_relative 'cache/active_support_mem_cache_store'
+  require_relative 'cache/active_support_dalli_store'
 end