+++ spec/support/spec_support/meta_scopes.rb
@@ -29,6 +29,10 @@ RSpec.configure do |config|
     example.call if test_as_memory_store_cache?
   end
 
+  config.around(:example, :as_dalli_store) do |example|
+    example.call if test_as_dalli_store_cache?
+  end
+
   # rubocop:disable Metrics/LineLength
   config.around(:example, :exclude) do |example|
     next if test_redis_cache?                && example.metadata[:exclude].include?(:redis)
@@ -38,6 +42,7 @@ RSpec.configure do |config|
     next if test_as_mem_cache_store_cache?   && example.metadata[:exclude].include?(:as_mem_cache_store)
     next if test_as_file_store_cache?        && example.metadata[:exclude].include?(:as_file_store)
     next if test_as_memory_store_cache?      && example.metadata[:exclude].include?(:as_memory_store)
+    next if test_as_dalli_store_cache?       && example.metadata[:exclude].include?(:as_dalli_store)
 
     example.call
   end
@@ -52,6 +57,7 @@ RSpec.configure do |config|
     next if test_as_mem_cache_store_cache?   && !example.metadata[:only].include?(:as_mem_cache_store)
     next if test_as_file_store_cache?        && !example.metadata[:only].include?(:as_file_store)
     next if test_as_memory_store_cache?      && !example.metadata[:only].include?(:as_memory_store)
+    next if test_as_dalli_store_cache?       && !example.metadata[:only].include?(:as_dalli_store)
 
     example.call
   end